fix php_compatibility error
