+++ modules/custom/met_api/src/Plugin/rest/ApiResource/EventReportResource.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Drupal\met_api\Plugin\rest\resource;
+namespace Drupal\met_api\Plugin\rest\ApiResource;
 
 use Drupal\Core\Session\AccountProxyInterface;
 use Drupal\Core\StringTranslation\StringTranslationTrait;
@@ -14,14 +14,14 @@ use Symfony\Component\DependencyInjection\ContainerInterface;
  * Provides the API resource for the mobile App.
  *
  * @RestResource(
- *   id = "met_api_impact_report_resource",
- *   label = @Translation("MET API Impact Report Resouce"),
+ *   id = "met_event_report_api_resource",
+ *   label = @Translation("MET API Event Report Resouce"),
  *   uri_paths = {
- *      "create" = "/api/v1/impact-report"
+ *      "create" = "/api/v1/event-report"
  *   }
  * )
  */
-class ImpactReportResource extends ResourceBase {
+class EventReportResource extends ResourceBase {
 
 
   use StringTranslationTrait;
@@ -83,7 +83,7 @@ class ImpactReportResource extends ResourceBase {
   public function post($data) {
 
     $response_code = 201;
-    $response_msg = 'Impact report API endpoint';
+    $response_msg = 'Event Report API endpoint';
 
     /*
     if (!$this->currentUser->hasPermission('administer site content')) {
@@ -100,83 +100,60 @@ class ImpactReportResource extends ResourceBase {
 
       $node = Node::create(
         [
-          'type' => 'impact_report',
-          'title' => 'Impact Report',
-          'field_full_name' => $value['full_name'],
-          'field_phone_number' => $value['phone'],
-          'field_location' => strtolower($value['location']),
-          'field_impact_category' => strtolower($value['category']),
+          'type' => 'event_report',
+          'title' => $value['title'],
           'body' => [
             'summary' => '',
             'value' => $value['body'],
             'format' => 'full_html',
           ],
-          'field_anyone_missing' => $value['anyone_missing'],
-          'field_anyone_passed_away' => $value['anyone_passed_away'],
-          'field_impacted_items' => $value['impacted_items'],
           'field_images' => $images,
-          'field_event' => $value['event_id'],
           'field_geo_location' => [
             'lat' => $value['lat'],
-            'lng' => $value['lon'],
+            'lng' => $value['lng'],
           ],
-          'field_village' => $value['village'],
         ]
       );
 
-      $node->enforceIsNew();
+      // Check permission.
       $check = $node->access('create', $this->currentUser);
 
       if (!$check) {
-        \Drupal::logger('MET API')->notice('Access denied, trying to create Impact Report');
+        \Drupal::logger('MET API')->notice('Access denied, trying to create Event Report');
         $response_msg = 'Access Denied.';
         $response_code = 403;
         return $this->response($response_msg, $response_code);
       }
 
+      $node->enforceIsNew();
       $node->save();
+      $node->access('create', $this->currentUser);
       $this->logger->notice($this->t("Node with nid @nid saved! \n", ['@nid' => $node->id()]));
       $nodes[] = $node->id();
     }
+
     $response_msg = $this->t("New Nodes creates with nids : @message", ['@message' => implode(",", $nodes)]);
 
     // Pass data to websocket server to deliver
     // ---------------------------------------------.
     $current_time = \Drupal::time()->getCurrentTime();
 
-    // Get village name from term id.
-    $term = \Drupal::service('entity_type.manager')->getStorage('taxonomy_term')->load($data[0]['village']);
-    $village = $term->getName();
-
-    // Get event name from event id.
-    $event = \Drupal::service('entity_type.manager')->getStorage('node')->load($data[0]['event_id']);
-    $event_name = $event->getTitle();
-
     $p = [
-      'name' => $data[0]['full_name'],
-      'phone' => $data[0]['phone'],
-      'location' => $data[0]['location'],
-      'category' => $data[0]['category'],
-      'missing' => $data[0]['anyone_missing'],
+      'title' => $data[0]['title'],
       'body' => $data[0]['body'],
-      'pass_away' => $data[0]['anyone_passed_away'],
-      'impacted_items' => $data[0]['impacted_items'],
-      'photo' => $data[0]['images'],
-      'event_id' => $event_name,
-      'lat' => $data[0]['lat'],
-      'lon' => $data[0]['lon'],
-      'village' => $village,
+      'lat' => "{$data[0]['lat']}",
+      'lon' => "{$data[0]['lng']}",
       'date' => date('d/m/Y', $current_time),
       'time' => date('h:i a', $current_time),
-      'type' => 'Impact Report',
-    // <-- unique id
-      'id' => $nodes[0] . 'ir',
+      'photo' => $data[0]['images'],
+      'type' => 'Event Report',
+      'id' => $nodes[0],
     ];
 
     $payload = [
       'action' => 'message',
       'username' => 'drupal',
-      'etype' => 'impact_report',
+      'etype' => 'met_event_report',
       'userrole' => 'tms',
       'payload' => $p,
     ];
@@ -194,7 +171,6 @@ class ImpactReportResource extends ResourceBase {
     $tms_socket_service->send($payload);
 
     return $this->response($response_msg, $response_code);
-
   }
 
   /**
@@ -209,7 +185,13 @@ class ImpactReportResource extends ResourceBase {
    *
    */
   public function permissions() {
-    return [];
+    return [
+      'MET API permission for event report' => [
+        'title' => $this->t('MET API permission for Event Report'),
+        'description' => $this->t('This is a permission to allow access to MET API event report'),
+        'restrict access' => TRUE,
+      ],
+    ];
   }
 
 }
