+++ modules/custom/met_api/src/Plugin/rest/ApiResource/EventResource.php
@@ -1,11 +1,10 @@
 <?php
 
-namespace Drupal\met_api\Plugin\rest\resource;
+namespace Drupal\met_api\Plugin\rest\ApiResource;
 
 use Drupal\Core\Cache\CacheableMetadata;
 use Drupal\Core\Session\AccountProxyInterface;
 use Drupal\Core\StringTranslation\StringTranslationTrait;
-use Drupal\image\Entity\ImageStyle;
 use Drupal\rest\Plugin\ResourceBase;
 use Drupal\rest\ResourceResponse;
 use Psr\Log\LoggerInterface;
@@ -15,14 +14,14 @@ use Symfony\Component\DependencyInjection\ContainerInterface;
  * Provides the API resource for the mobile App.
  *
  * @RestResource(
- *   id = "met_api_evacuation_resource",
- *   label = @Translation("MET API Evacuation Resouce"),
+ *   id = "met_api_event_resource",
+ *   label = @Translation("MET API Event Resouce"),
  *   uri_paths = {
- *      "canonical" = "/api/v1/evacuation"
+ *      "canonical" = "/api/v1/event"
  *   }
  * )
  */
-class EvacuationResource extends ResourceBase {
+class EventResource extends ResourceBase {
 
   use StringTranslationTrait;
 
@@ -66,7 +65,7 @@ class EvacuationResource extends ResourceBase {
       $plugin_definition,
       $container->getParameter('serializer.formats'),
       $container->get('logger.factory')->get('MET API'),
-      $container->get('current_user'),
+      $container->get('current_user')
     );
   }
 
@@ -77,7 +76,8 @@ class EvacuationResource extends ResourceBase {
 
     $storage = \Drupal::service('entity_type.manager')->getStorage('node');
     $nids = $storage->getQuery()
-      ->condition('type', 'evacuation')
+      ->condition('type', 'event')
+      ->condition('field_active', 1)
       ->condition('status', 1)
       ->sort('created', 'DESC')
       ->accessCheck(FALSE)
@@ -85,35 +85,66 @@ class EvacuationResource extends ResourceBase {
 
     $nodes = $storage->loadMultiple($nids);
     $new_nodes = [];
+    $paragraph_fields_include = ['type', 'status', 'field_date', 'field_date_only', 'field_region_', 'field_location', 'field_depth', 'field_magnitude', 'field_category', 'field_name', 'field_active'];
     foreach ($nodes as $node) {
-
-      // Process location.
-      $lat = '';
-      $lon = '';
-      if ($node->field_geo_location->value != "") {
-        [$lat, $lon] = explode(", ", $node->field_geo_location->value);
-      }
-
-      // Process photo
-      // $large_image = ImageStyle::load('large')->buildUrl($node->field_safe_zone_image->entity->getFileUri());
-      $large_image = \Drupal::service('file_url_generator')->generateAbsoluteString($node->field_safe_zone_image->entity->getFileUri());
-      $thumb_image = ImageStyle::load('thumbnail')->buildUrl($node->field_safe_zone_image->entity->getFileUri());
-
       $data = [];
-      $data['id'] = (int) $node->id();
+      $data['id'] = $node->id();
       $data['title'] = $node->title->value;
       $data['body'] = $node->body->value != '' ? strip_tags($node->body->value) : $node->body->value;
-      $data['image_large'] = $large_image;
-      $data['image_small'] = $thumb_image;
-      $data['lat'] = (Double) $lat;
-      $data['lon'] = (Double) $lon;
+      $data['active'] = $node->field_active->value;
 
-      $new_nodes[] = $data;
+      $fields = $node->field_event_type->referencedEntities();
+      foreach ($paragraph_fields_include as $field) {
+        if (isset($fields[0]->$field)) {
+          if ($field == 'type') {
+            $data[$field] = $fields[0]->$field->target_id;
+          }
+          elseif ($field == 'field_date') {
+            $data['date'] = $fields[0]->$field->date->format('d/m/Y');
+            $data['time'] = $fields[0]->$field->date->format('h:i a');
+            $data['field_date'] = $fields[0]->$field->date->format('d/m/Y h:i a');
+          }
+          elseif ($field == 'field_date_only') {
+            $data['date'] = $fields[0]->$field->date->format('d/m/Y');
+          }
+          else {
+            $data[$field] = $fields[0]->$field->value;
+          }
+        }
+      }
+
+      // Check to see if any data in feel earthquake for this event.
+      if ($data['type'] == 'earthquake') {
+        $str = \Drupal::service('entity_type.manager')->getStorage('met_feel_earthquake');
+        $itemids = $str->getQuery()
+          ->condition('field_event', $node->id())
+          ->accessCheck(FALSE)
+          ->addTag('debug')
+          ->execute();
+
+        $feels = $str->loadMultiple($itemids);
+        $location = [];
+        foreach ($feels as $field) {
+          $location[] = $field->field_location->value;
+        }
+
+        $uLocation = array_unique($location);
+        $feel_data = [];
+        foreach ($uLocation as $value) {
+          $feel_data[$value] = count(array_intersect($location, [$value]));
+        }
+
+        if (count($feel_data) > 0) {
+          $data['feel'] = $feel_data;
+        }
+      }
+
+      $new_nodes[$node->id()] = $data;
     }
 
     $build = [
       '#cache' => [
-        'tags' => ['node_list:evacuation'],
+        'tags' => ['node_list:event', 'met_feel_earthquake_list'],
       ],
     ];
