+++ modules/custom/met_api/src/Plugin/rest/ApiResource/ImpactReportResource.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Drupal\met_api\Plugin\rest\resource;
+namespace Drupal\met_api\Plugin\rest\ApiResource;
 
 use Drupal\Core\Session\AccountProxyInterface;
 use Drupal\Core\StringTranslation\StringTranslationTrait;
@@ -14,14 +14,14 @@ use Symfony\Component\DependencyInjection\ContainerInterface;
  * Provides the API resource for the mobile App.
  *
  * @RestResource(
- *   id = "met_api_request_assistance_resource",
- *   label = @Translation("MET API Request Assistance Resouce"),
+ *   id = "met_api_impact_report_resource",
+ *   label = @Translation("MET API Impact Report Resouce"),
  *   uri_paths = {
- *      "create" = "/api/v1/request-assistance"
+ *      "create" = "/api/v1/impact-report"
  *   }
  * )
  */
-class RequestAssistanceResource extends ResourceBase {
+class ImpactReportResource extends ResourceBase {
 
 
   use StringTranslationTrait;
@@ -83,7 +83,7 @@ class RequestAssistanceResource extends ResourceBase {
   public function post($data) {
 
     $response_code = 201;
-    $response_msg = 'Request assistance API endpoint';
+    $response_msg = 'Impact report API endpoint';
 
     /*
     if (!$this->currentUser->hasPermission('administer site content')) {
@@ -100,24 +100,22 @@ class RequestAssistanceResource extends ResourceBase {
 
       $node = Node::create(
         [
-          'type' => 'request_assistance',
-          'title' => 'Request Assistance',
+          'type' => 'impact_report',
+          'title' => 'Impact Report',
           'field_full_name' => $value['full_name'],
           'field_phone_number' => $value['phone'],
           'field_location' => strtolower($value['location']),
+          'field_impact_category' => strtolower($value['category']),
           'body' => [
             'summary' => '',
             'value' => $value['body'],
             'format' => 'full_html',
           ],
-
-          'field_assistance_with' => $value['assistance_with'],
-          'field_needed_now' => $value['needed_now'],
+          'field_anyone_missing' => $value['anyone_missing'],
+          'field_anyone_passed_away' => $value['anyone_passed_away'],
+          'field_impacted_items' => $value['impacted_items'],
           'field_images' => $images,
           'field_event' => $value['event_id'],
-          'field_have_water' => $value['have_water'],
-          'field_have_house_shelter' => $value['have_house'],
-          'field_have_food' => $value['have_food'],
           'field_geo_location' => [
             'lat' => $value['lat'],
             'lng' => $value['lon'],
@@ -126,23 +124,20 @@ class RequestAssistanceResource extends ResourceBase {
         ]
       );
 
-      // Check access permission.
+      $node->enforceIsNew();
       $check = $node->access('create', $this->currentUser);
 
       if (!$check) {
-        \Drupal::logger('MET API')->notice('Access denied, trying to create Assistance Report');
+        \Drupal::logger('MET API')->notice('Access denied, trying to create Impact Report');
         $response_msg = 'Access Denied.';
         $response_code = 403;
         return $this->response($response_msg, $response_code);
       }
 
-      $node->enforceIsNew();
       $node->save();
-      $node->access('create', $this->currentUser);
       $this->logger->notice($this->t("Node with nid @nid saved! \n", ['@nid' => $node->id()]));
       $nodes[] = $node->id();
     }
-
     $response_msg = $this->t("New Nodes creates with nids : @message", ['@message' => implode(",", $nodes)]);
 
     // Pass data to websocket server to deliver
@@ -161,12 +156,11 @@ class RequestAssistanceResource extends ResourceBase {
       'name' => $data[0]['full_name'],
       'phone' => $data[0]['phone'],
       'location' => $data[0]['location'],
-      'have_shelter' => $data[0]['have_house'],
-      'assistance_with' => $data[0]['assistance_with'],
+      'category' => $data[0]['category'],
+      'missing' => $data[0]['anyone_missing'],
       'body' => $data[0]['body'],
-      'needed_now' => $data[0]['needed_now'],
-      'have_water' => $data[0]['have_water'],
-      'have_food' => $data[0]['have_food'],
+      'pass_away' => $data[0]['anyone_passed_away'],
+      'impacted_items' => $data[0]['impacted_items'],
       'photo' => $data[0]['images'],
       'event_id' => $event_name,
       'lat' => $data[0]['lat'],
@@ -174,15 +168,15 @@ class RequestAssistanceResource extends ResourceBase {
       'village' => $village,
       'date' => date('d/m/Y', $current_time),
       'time' => date('h:i a', $current_time),
-      'type' => 'Request Assistance',
+      'type' => 'Impact Report',
     // <-- unique id
-      'id' => $nodes[0] . 'ra',
+      'id' => $nodes[0] . 'ir',
     ];
 
     $payload = [
       'action' => 'message',
       'username' => 'drupal',
-      'etype' => 'request_assistance',
+      'etype' => 'impact_report',
       'userrole' => 'tms',
       'payload' => $p,
     ];
@@ -200,6 +194,7 @@ class RequestAssistanceResource extends ResourceBase {
     $tms_socket_service->send($payload);
 
     return $this->response($response_msg, $response_code);
+
   }
 
   /**
