+++ modules/custom/met_api/src/Plugin/rest/ApiResource/VillageResource.php
@@ -1,7 +1,8 @@
 <?php
 
-namespace Drupal\met_api\Plugin\rest\resource;
+namespace Drupal\met_api\Plugin\rest\ApiResource;
 
+use Drupal\Core\Cache\CacheableMetadata;
 use Drupal\Core\Session\AccountProxyInterface;
 use Drupal\Core\StringTranslation\StringTranslationTrait;
 use Drupal\rest\Plugin\ResourceBase;
@@ -13,14 +14,15 @@ use Symfony\Component\DependencyInjection\ContainerInterface;
  * Provides the API resource for the mobile App.
  *
  * @RestResource(
- *   id = "met_api_notification_resource",
- *   label = @Translation("MET API Notification Resouce"),
+ *   id = "met_api_village_resource",
+ *   label = @Translation("MET API Village Resouce"),
  *   uri_paths = {
- *      "canonical" = "/api/v1/notification/{lng}"
+ *      "canonical" = "/api/v1/village/{rid}"
  *   }
  * )
  */
-class NotificationResource extends ResourceBase {
+class VillageResource extends ResourceBase {
+
   use StringTranslationTrait;
 
   /**
@@ -70,45 +72,29 @@ class NotificationResource extends ResourceBase {
   /**
    *
    */
-  public function get($lng = 'en') {
+  public function get($rid) {
+    $vocabulary = "cities";
+    $child_terms = \Drupal::service('entity_type.manager')->getStorage('taxonomy_term')->loadTree(
+      $vocabulary,
+      $rid,
+      NULL,
+      FALSE);
 
-    $storage = \Drupal::service('entity_type.manager')->getStorage('met_notification');
-    $items = $storage->getQuery()
-      ->condition('status', 1)
-      ->accessCheck(FALSE)
-      ->sort('created', 'DESC')
-      ->range(0, 10)
-      ->execute();
-
-    $items = $storage->loadMultiple($items);
-    $new_items = [];
-    foreach ($items as $item) {
-
-      $data = [];
-      foreach ($item->field_language as $p) {
-        $ent = $p->entity;
-        if ($lng == 'en' && $ent->type->target_id == 'notification_english') {
-          $data['body'] = strip_tags($ent->field_body->value);
-          $data['title'] = strip_tags($ent->field_title->value);
-        }
-        if ($lng == 'to' && $ent->type->target_id == 'notification_tongan') {
-          $data['body'] = strip_tags($ent->field_body->value);
-          $data['title'] = strip_tags($ent->field_title->value);
-        }
-      }
-
-      $data['id'] = $item->id();
-      $data['level'] = $item->field_level->value;
-      $data['target_location'] = $item->field_location;
-      $data['time'] = \Drupal::service('date.formatter')->format($item->created->value, 'custom', 'd/m/Y');
-      $data['date'] = \Drupal::service('date.formatter')->format($item->created->value, 'custom', 'h:i a');
-      $data['timestamp'] = $item->created->value;
-
-      $new_items[$item->id()] = $data;
+    $data = [];
+    foreach ($child_terms as $term) {
+      $output = [];
+      $output['id'] = $term->tid;
+      $output['name'] = $term->name;
+      $data[] = $output;
     }
-    $build = ['#cache' => ['max-age' => 0]];
 
-    return (new ResourceResponse($new_items, 200))->addCacheableDependency($build);
+    $build = [
+      '#cache' => [
+        'tags' => ['term_list:cities'],
+      ],
+    ];
+
+    return (new ResourceResponse($data, 200))->addCacheableDependency(CacheableMetadata::createFromRenderArray($build));
   }
 
   /**
